% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eighteen.R
\name{eighteen_la}
\alias{eighteen_la}
\title{Compete with Sausage vendor}
\usage{
eighteen_la(scoreI, d)
}
\arguments{
\item{scoreI}{Interger, the score you get in play_eighteen_la function}

\item{d}{difficulty, which range is 1~10}
}
\value{
It will tell you are win or lose or tie, and tell you the vendor's score and dices points.
}
\description{
Now you compete with Sausage vendor(Computer)!!
Enter the score and challenge with different difficulty!!
}
\examples{
sausage(throw)->dice
play_eighteen_la(dice)->score
eighteen_la(score[1],1)
}
