% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eighteen.R
\name{play_eighteen_la}
\alias{play_eighteen_la}
\title{Calculate the score by rule}
\usage{
play_eighteen_la(dices)
}
\arguments{
\item{dices}{The result get by sausage function(sampling), it must be the vector form.}
}
\value{
If the player doesn't play the second time, the function will return only score.
If the player play the second time, it will return both score and the dices' points we get in the sceond time with array, and the first position of the return is score and another is points.
}
\description{
After throwing a dice, we can calculate the score by rule.
If the score we get in first time is zero, we can throw one more time.If we throw the second time but the score still be zero, the score will be zero.
}
\examples{
##No second times
dices=c(2,2,5,6)
play_eighteen_la(dices)
##second times
dices=c(1,1,1,4)
play_eighteen_la(dices)
}
