% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_charming_fixed.r
\name{play_showhand}
\alias{play_showhand}
\title{Play showhand with multiple persons}
\usage{
play_showhand(people)
}
\arguments{
\item{people}{person you want to choose.}
}
\value{
The result is data frame, and V6 is the cards type.
}
\description{
In this function, you can throw it with many times, and it will tell you the result.
}
\details{
the result type(V6) of showhand, it appears 2~9
2 is High card
3 is one pair
4 is two pairs
5 is Three of a kind
6 is Straight
7 is Full house
8 is Four of a Kind
9 is Five of a kind
}
\examples{
###############
play_showhand(10)
##############You can also see who wins the game######
play_showhand(10)->result
showHand_com(result)

}
