% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eighteen.R
\name{sausage}
\alias{sausage}
\title{Tranditional taiwanese game at sausage street vendor}
\usage{
sausage(throw)
}
\arguments{
\item{throw}{type everything can execute this function, also NULL.}
}
\value{
equivlent to the result of throwing four dices.
}
\description{
Function for tranditional taiwanese game which usually appears in sausage street vendor in Taiwan.
This function can throw a dice with four dice.
}
\examples{
##Both are equivalent
sausage(throw)
sample(c(1:6),4,replace = TRUE)
}
