% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_charming_fixed.r
\name{showhand}
\alias{showhand}
\title{Type of Showhand game in dices}
\usage{
showhand(y)
}
\arguments{
\item{y}{a result of throwing dices (sampling)}
}
\value{
type of showhand
}
\description{
Showhand game is very popular game in casino, this function can classify the type of showhand with dices.We throw five dices and watch the type.
}
\details{
the result is type of showhand, it appears 2~9.
2 is High card ||
3 is one pair ||
4 is two pairs ||
5 is Three of a kind ||
6 is Straight ||
7 is Full house ||
8 is Four of a Kind ||
9 is Five of a kind ||
}
\examples{
##########
show=sample(c(1:6),5,replace = TRUE)
show
showhand(show)

}
