% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_charming_fixed.r
\name{showHand_com}
\alias{showHand_com}
\title{Multiple players play Showhand with dices}
\usage{
showHand_com(result)
}
\arguments{
\item{result}{The result we get in function play_showhand, which is data.frame type.}
}
\value{
The original data frame and tell you who is/are  win in the game! The sixth column is the dices type, it is chinese type.
}
\description{
We can put the result we get in function play_showhand into the function, and determine who is win!
}
\examples{
#################
play_showhand(5)->play
showHand_com(play)
}
